+++ Model/Resources/Strings.Designer.cs
@@ -473,7 +473,7 @@ namespace Tellurian.Trains.Schedules.Model.Resources {
                 return ResourceManager.GetString("TrainHasNoLocomotiveAssigned", resourceCulture);
             }
         }
-
+        
         /// <summary>
         ///   Looks up a localized string similar to Train {0} runs from {1} {2} to {3} {4}, but the layout has no track stretch between these locations..
         /// </summary>
@@ -482,9 +482,9 @@ namespace Tellurian.Trains.Schedules.Model.Resources {
                 return ResourceManager.GetString("TrainHasNoStretchBetweenCalls", resourceCulture);
             }
         }
-
+        
         /// <summary>
-        ///   Looks up a localized string similar to Train {0} has no traction unit on sessions {1}..
+        ///   Looks up a localized string similar to Train {0} has no traction unit between {1} and {2} on sessions {3}..
         /// </summary>
         internal static string TrainMissingTraction {
             get {